refactor: :fire: refactor constants in consts.rs file and reduce load when drawing stuff

+++ src/samples.rs
@@ -4,20 +4,20 @@ use egui::plot::{Value, Values};
 
 #[derive(Debug)]
 pub struct Samples {
-    max_samples: usize,
+    max_samples: u32,
     inner: VecDeque<Value>,
 }
 
 impl Samples {
-    pub fn new(max_samples: usize) -> Self {
+    pub fn new(max_samples: u32) -> Self {
         Samples {
             max_samples,
-            inner: VecDeque::with_capacity(max_samples),
+            inner: VecDeque::with_capacity(max_samples as usize),
         }
     }
 
     pub fn insert(&mut self, sample: Value) {
-        if self.inner.len() == self.max_samples {
+        if self.inner.len() as u32 == self.max_samples {
             self.inner.pop_front().unwrap();
         }
